4 recording the bare minimum (#15)
* [4] Recording bare minimum

+++ protocol-amqp-091/src/main/java/org/kendar/amqp/v09/AmqpFileStorage.java
@@ -4,23 +4,33 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.kendar.amqp.v09.utils.AmqpStorage;
 import org.kendar.storage.BaseFileStorage;
+import org.kendar.storage.CompactLine;
 import org.kendar.storage.StorageItem;
+import org.kendar.utils.JsonMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> implements AmqpStorage {
     private static final Logger log = LoggerFactory.getLogger(AmqpFileStorage.class);
+    private static final List<String> toAvoid = List.of("byte[]",
+            "ConnectionStartOk", "ConnectionTuneOk", "ConnectionOpen", "ChannelOpen", "BasicPublish",
+            "HeaderFrame", "BasicPublish", "BodyFrame", "BasicAck", "ChannelClose", "ConnectionClose",
+            "QueueDeclare", "ExchangeDeclare", "QueueDelete", "ExchangeDelete");
     private final List<StorageItem<JsonNode, JsonNode>> inMemoryDb = new ArrayList<>();
     private final List<StorageItem<JsonNode, JsonNode>> compareData = new ArrayList<>();
     private final List<StorageItem<JsonNode, JsonNode>> outItems = new ArrayList<>();
     private final Object lockObject = new Object();
     private final Object responseLockObject = new Object();
     private boolean initialized = false;
+    private List<CompactLine> index;
+    private final JsonMapper mapper = new JsonMapper();
 
     public AmqpFileStorage(String targetDir) {
         super(targetDir);
@@ -30,10 +40,29 @@ public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> impleme
         super(targetDir);
     }
 
+    private static int getConsumeId(JsonNode output, int consumeId) {
+        if (output == null) return 0;
+        var data = output.get("data");
+        if (data == null) return consumeId;
+        var cid = data.get("consumeId");
+        if (cid == null) return consumeId;
+        return Math.max(cid.asInt(), consumeId);
+    }
+
     @Override
     protected TypeReference<?> getTypeReference() {
         return new TypeReference<StorageItem<JsonNode, JsonNode>>() {
         };
+    }
+
+    @Override
+    protected boolean shouldNotSave(CompactLine cl, List<CompactLine> compactLines, StorageItem<JsonNode, JsonNode> item, List<StorageItem<JsonNode, JsonNode>> loadedData) {
+        if (useFullData) return false;
+        if (cl == null) return false;
+        if (cl.getTags() == null || cl.getTags().get("input") == null) {
+            return false;
+        }
+        return toAvoid.contains(cl.getTags().get("input"));
     }
 
     @Override
@@ -43,10 +72,33 @@ public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> impleme
             var item = inMemoryDb.stream()
                     .filter(a -> type.equalsIgnoreCase(a.getType()) &&
                             a.getCaller().equalsIgnoreCase("AMQP")).findFirst();
-            if (item.isPresent()) {
+            var idx = index.stream()
+                    .filter(a -> type.equalsIgnoreCase(a.getType()) &&
+                            a.getCaller().equalsIgnoreCase("AMQP")).findFirst();
+
+            CompactLine cl = null;
+            if (idx.isPresent()) {
+                cl = idx.get();
+            }
+            var shouldNotSave = shouldNotSave(cl, null, null, null);
+            if (item.isPresent() && !shouldNotSave) {
+
+                log.debug("[SERVER][REPFULL]  " + item.get().getIndex() + ":" + item.get().getType());
                 inMemoryDb.remove(item.get());
+                if (idx.isPresent()) {
+                    index.remove(idx.get());
+                }
                 return item.get();
             }
+
+            if (idx.isPresent()) {
+                log.debug("[SERVER][REPSHRT] " + idx.get().getIndex() + ":" + idx.get().getType());
+                index.remove(idx.get());
+                var sti = new StorageItem<JsonNode, JsonNode>();
+                sti.setIndex(idx.get().getIndex());
+                return sti;
+            }
+
 
             return null;
         }
@@ -62,6 +114,8 @@ public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> impleme
                 }
                 inMemoryDb.add(item);
             }
+
+            index = retrieveIndexFile();
             initialized = true;
         }
     }
@@ -71,12 +125,14 @@ public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> impleme
         initializeContent();
         synchronized (responseLockObject) {
             var result = new ArrayList<StorageItem<JsonNode, JsonNode>>();
-            for (var item : compareData.stream().filter(a -> a.getIndex() > afterIndex).collect(Collectors.toList())) {
+            for (var item : index.stream().filter(a -> a.getIndex() > afterIndex).collect(Collectors.toList())) {
                 if (item.getType().equalsIgnoreCase("RESPONSE")) {
-                    if (outItems.contains(item)) {
-                        result.add(item);
-                        log.trace("[SERVER][CB] After: " + afterIndex + " Index: " + item.getIndex() + " Type: " + item.getOutput().get("type").textValue());
-                        outItems.remove(item);
+                    var outItem = outItems.stream().filter(a -> a.getIndex() == item.getIndex()).findFirst();
+                    if (!outItem.isEmpty()) {
+                        result.add(outItem.get());
+                        log.debug("[SERVER][CB] After: " + afterIndex + " Index: " + item.getIndex() + " Type: " +
+                                outItem.get().getOutput().get("type").textValue());
+                        outItems.remove(outItem.get());
                     }
                 } else {
                     break;
@@ -85,4 +141,35 @@ public class AmqpFileStorage extends BaseFileStorage<JsonNode, JsonNode> impleme
             return result;
         }
     }
+
+    protected Map<String, String> buildTag(StorageItem<JsonNode, JsonNode> item) {
+        var data = new HashMap<String, String>();
+        var consumeId = 0;
+        data.put("input", null);
+        data.put("output", null);
+        data.put("consumeId", null);
+        if (item.getInput() != null) {
+            if (item.getInput().get("type") != null) {
+                data.put("input", item.getInput().get("type").textValue());
+                consumeId = getConsumeId(item.getOutput(), consumeId);
+            }
+        }
+        if (item.getOutput() != null) {
+            if (item.getOutput().get("type") != null) {
+                data.put("output", item.getOutput().get("type").textValue());
+                consumeId = getConsumeId(item.getOutput(), consumeId);
+            }
+        }
+        if (consumeId > 0) {
+            data.put("consumeId", consumeId + "");
+        }
+        return data;
+    }
+
+    @Override
+    protected List<StorageItem<JsonNode, JsonNode>> readAllItems() {
+        var res = super.readAllItems();
+        return res;
+    }
+
 }
